move RenameFile back to org.openrewrite, add core category descriptor
