+++ rewrite-core/src/main/java/org/openrewrite/RenameFile.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java;
+package org.openrewrite;
 
 import lombok.EqualsAndHashCode;
 import lombok.Value;
